--- src/decorator/cookies.decorator.ts
@@ -1,10 +1,0 @@
-import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-
-export const Cookies = createParamDecorator(
-  (data: string, ctx: ExecutionContext) => {
-    const request = ctx.switchToHttp().getRequest();
-    return data ? request.cookies?.[data] : request.cookies;
-  },
-);
-
-export default Cookies;
